Introduce a `cli-helpers` module along with an initial command
The purpose of this new module is to provide a command line interface
offering helper commands for working with ORT such as e.g. working with
global or local ORT configuration files and OrtResult files.

It is important to note that ORT configuration files are likely to be
put under version control. This should always be considered when adding
new commands the output of commands which do several things at once would
lead to non-atomic commits.

In order to keep things atomic and to illustrate the approach this change
adds just one tiny command for re-formatting a repository configuration file
or `.ort.yml` as we name it.

Signed-off-by: Frank Viernau <frank.viernau@here.com>