reporter-web-app: Use Kotlin's Yarn plugin to bootstrap Node.js and Yarn
This allows to remove Node.js and Yarn from the list of prerequisites
to build ORT, which means now nothing else than a JDK is needed.

Unfortunately, the Yarn plugin always installs itself to the root project
(even if applied in a subproject) [1], and it somehow also creates a
"check" task in the root project which clashes with our existing "check"
task. As for some reason the "check" task created by the Yarn plugin
cannot be further configured to depend on our "checkCopyright" task,
simply drop our "check" task and call "checkCopyright" explicitly in CI
configurations.

[1] https://github.com/JetBrains/kotlin/blob/20626e4aaf918803aeb3c43257100bc95f63e45f/libraries/tools/kotlin-gradle-plugin/src/main/kotlin/org/jetbrains/kotlin/gradle/targets/js/yarn/YarnPlugin.kt#L27-L31

Signed-off-by: Sebastian Schuberth <sebastian.schuberth@bosch-si.com>