FindingsMatcherTest: Remove the randomization again
The reason why those two functions were originally put in place was
to make the calling code more compact by using reasonable default
values. As the majority of test cases using it is perfectly fine
with non-random default values it makes sense to remove that again.
One major reason why random values are not desired is because it makes
the tests non-deterministic and thus can introduce flakyness.

The random values have been introduced for communicating to the reader
that a value can be arbitrary. It is however questionable whether most
code readers actually interpret it this way. Moreover it is easy to introduce
random / flaky new tests or miss it in the code review when the function
is called using default parameter values.

Making it more clear which parameters are important for the test and which
one can be arbitrary is better done using a descriptive test name and thus
implicitly also by fixing other test smells like testing multiple things
at a time and so on.

This change is a partial revert of 4a5ebd81.

Signed-off-by: Frank Viernau <frank.viernau@here.com>